third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/correct-services.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices('heart_rate'))
    .then(services => Promise.all([
      services[0].getCharacteristics(),
      services[1].getCharacteristics()]))
    .then(characteristics_arrays => {
      assert_equals(characteristics_arrays[0].length, 2);
      assert_equals(characteristics_arrays[1].length, 1);
    });
}, 'Find correct services with UUID.');
</script>
